+++ src/table/apps.py
@@ -4,5 +4,5 @@ from __future__ import unicode_literals, print_function
 from django.apps import AppConfig
 
 
-class MatrixConfig(AppConfig):
-    name = 'matrix'
+class TableConfig(AppConfig):
+    name = 'table'