+++ src/workflow/models.py
@@ -13,7 +13,7 @@ from dataops import pandas_db
 
 class Workflow(models.Model):
     """
-    Model for a workflow, that is, a matrix, set of column descriptions and
+    Model for a workflow, that is, a table, set of column descriptions and
     all the information regarding the actions, conditions and such. This is
     the main object in the relational model.
     """